requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
python-dotenv>=1.0
jinja2>=3.1
openai>=1.0
